requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
streamlit==1.38.0
pydantic==2.9.2
python-multipart==0.0.9
httpx==0.27.2

# Core scientific stack
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
matplotlib==3.9.2
scikit-learn==1.5.2

# Audio processing
librosa==0.10.2.post1
soundfile==0.12.1

# PyTorch (CPU) and vision
torch==2.4.1
torchvision==0.19.1

# Geo utilities
geopy==2.4.1
timm==1.0.9
